Keep only "realistic" RLC UM (now it is called LteRlcUm/LteRlcUmD2D)
