+++ src/stack/mac/layer/LteMacUeRealistic.cc
@@ -42,8 +42,8 @@ void LteMacUeRealistic::initialize(int stage)
         // TODO do the same for RLC AM
         std::string rlcUmType = getParentModule()->getSubmodule("rlc")->par("LteRlcUmType").stdstringValue();
         std::string macType = getParentModule()->par("LteMacType").stdstringValue();
-        if (macType.compare("LteMacUeRealistic") == 0 &&  rlcUmType.compare("LteRlcUmRealistic") != 0)
-            throw cRuntimeError("LteMacUeRealistic::initialize - %s module found, must be LteRlcUmRealistic. Aborting", rlcUmType.c_str());
+        if (macType.compare("LteMacUeRealistic") == 0 &&  rlcUmType.compare("LteRlcUm") != 0)
+            throw cRuntimeError("LteMacUeRealistic::initialize - %s module found, must be LteRlcUm. Aborting", rlcUmType.c_str());
     }
 }
