+++ src/stack/rlc/um/LteRlcUm.cc
@@ -8,31 +8,39 @@
 //
 
 #include "stack/rlc/um/LteRlcUm.h"
-#include "stack/d2dModeSelection/D2DModeSwitchNotification_m.h"
+#include "stack/mac/packet/LteMacSduRequest.h"
 
 Define_Module(LteRlcUm);
 
-UmTxQueue* LteRlcUm::getTxBuffer(MacNodeId nodeId, LogicalCid lcid)
+UmTxEntity* LteRlcUm::getTxBuffer(FlowControlInfo* lteInfo)
 {
+    MacNodeId nodeId = ctrlInfoToUeId(lteInfo);
+    LogicalCid lcid = lteInfo->getLcid();
+
     // Find TXBuffer for this CID
     MacCid cid = idToMacCid(nodeId, lcid);
-    UmTxBuffers::iterator it = txBuffers_.find(cid);
-    if (it == txBuffers_.end())
+    UmTxEntities::iterator it = txEntities_.find(cid);
+    if (it == txEntities_.end())
     {
         // Not found: create
         std::stringstream buf;
         // FIXME HERE
 
-        buf << "UmTxQueue Lcid: " << lcid;
-        cModuleType* moduleType = cModuleType::get("lte.stack.rlc.UmTxQueue");
-        UmTxQueue* txbuf = check_and_cast<UmTxQueue *>(
-            moduleType->createScheduleInit(buf.str().c_str(), getParentModule()));
-        txBuffers_[cid] = txbuf;    // Add to tx_buffers map
+        buf << "UmTxEntity Lcid: " << lcid;
+        cModuleType* moduleType = cModuleType::get("lte.stack.rlc.UmTxEntity");
+        UmTxEntity* txEnt = check_and_cast<UmTxEntity *>(moduleType->createScheduleInit(buf.str().c_str(), getParentModule()));
+        txEntities_[cid] = txEnt;    // Add to tx_entities map
 
-        EV << "LteRlcUm : Added new UmTxBuffer: " << txbuf->getId() <<
+        if (lteInfo != NULL)
+        {
+            // store control info for this flow
+            txEnt->setFlowControlInfo(lteInfo->dup());
+        }
+
+        EV << "LteRlcUm : Added new UmTxEntity: " << txEnt->getId() <<
         " for node: " << nodeId << " for Lcid: " << lcid << "\n";
 
-        return txbuf;
+        return txEnt;
     }
     else
     {
@@ -44,30 +52,42 @@ UmTxQueue* LteRlcUm::getTxBuffer(MacNodeId nodeId, LogicalCid lcid)
     }
 }
 
-UmRxQueue* LteRlcUm::getRxBuffer(MacNodeId nodeId, LogicalCid lcid)
+
+UmRxEntity* LteRlcUm::getRxBuffer(FlowControlInfo* lteInfo)
 {
+    MacNodeId nodeId;
+    if (lteInfo->getDirection() == DL)
+        nodeId = lteInfo->getDestId();
+    else
+        nodeId = lteInfo->getSourceId();
+    LogicalCid lcid = lteInfo->getLcid();
+
     // Find RXBuffer for this CID
     MacCid cid = idToMacCid(nodeId, lcid);
-    UmRxBuffers::iterator it = rxBuffers_.find(cid);
-    if (it == rxBuffers_.end())
+
+    UmRxEntities::iterator it = rxEntities_.find(cid);
+    if (it == rxEntities_.end())
     {
         // Not found: create
         std::stringstream buf;
-        buf << "UmRxQueue Lcid: " << lcid;
-        cModuleType* moduleType = cModuleType::get("lte.stack.rlc.UmRxQueue");
-        UmRxQueue* rxbuf = check_and_cast<UmRxQueue *>(
+        buf << "UmRxEntity Lcid: " << lcid;
+        cModuleType* moduleType = cModuleType::get("lte.stack.rlc.UmRxEntity");
+        UmRxEntity* rxEnt = check_and_cast<UmRxEntity *>(
             moduleType->createScheduleInit(buf.str().c_str(), getParentModule()));
-        rxBuffers_[cid] = rxbuf;    // Add to rx_buffers map
+        rxEntities_[cid] = rxEnt;    // Add to rx_entities map
 
-        EV << "LteRlcUm : Added new UmRxBuffer: " << rxbuf->getId() <<
+        // store control info for this flow
+        rxEnt->setFlowControlInfo(lteInfo->dup());
+
+        EV << "LteRlcUm : Added new UmRxEntity: " << rxEnt->getId() <<
         " for node: " << nodeId << " for Lcid: " << lcid << "\n";
 
-        return rxbuf;
+        return rxEnt;
     }
     else
     {
         // Found
-        EV << "LteRlcUm : Using old UmRxBuffer: " << it->second->getId() <<
+        EV << "LteRlcUm : Using old UmRxEntity: " << it->second->getId() <<
         " for node: " << nodeId << " for Lcid: " << lcid << "\n";
 
         return it->second;
@@ -76,7 +96,7 @@ UmRxQueue* LteRlcUm::getRxBuffer(MacNodeId nodeId, LogicalCid lcid)
 
 void LteRlcUm::sendDefragmented(cPacket *pkt)
 {
-    Enter_Method("sendDefragmented()");                            // Direct Method Call
+    Enter_Method_Silent("sendDefragmented()");                    // Direct Method Call
     take(pkt);                                                    // Take ownership
 
     EV << "LteRlcUm : Sending packet " << pkt->getName() << " to port UM_Sap_up$o\n";
@@ -85,7 +105,7 @@ void LteRlcUm::sendDefragmented(cPacket *pkt)
 
 void LteRlcUm::sendFragmented(cPacket *pkt)
 {
-    Enter_Method("sendFragmented()");                            // Direct Method Call
+    Enter_Method_Silent("sendFragmented()");                      // Direct Method Call
     take(pkt);                                                    // Take ownership
 
     EV << "LteRlcUm : Sending packet " << pkt->getName() << " to port UM_Sap_down$o\n";
@@ -94,7 +114,7 @@ void LteRlcUm::sendFragmented(cPacket *pkt)
 
 void LteRlcUm::sendToLowerLayer(cPacket *pkt)
 {
-    Enter_Method("sendToLowerLayer()");                            // Direct Method Call
+    Enter_Method_Silent("sendToLowerLayer()");                    // Direct Method Call
     take(pkt);                                                    // Take ownership
     EV << "LteRlcUm : Sending packet " << pkt->getName() << " to port UM_Sap_down$o\n";
     send(pkt, down_[OUT]);
@@ -102,69 +122,103 @@ void LteRlcUm::sendToLowerLayer(cPacket *pkt)
 
 void LteRlcUm::handleUpperMessage(cPacket *pkt)
 {
+    EV << "LteRlcUm::handleUpperMessage - Received packet " << pkt->getName() << " from upper layer, size " << pkt->getByteLength() << "\n";
+
     FlowControlInfo* lteInfo = check_and_cast<FlowControlInfo*>(pkt->removeControlInfo());
-//    UmTxQueue* txbuf = getTxBuffer(lteInfo->getDestId(),lteInfo->getLcid());
-    UmTxQueue* txbuf = getTxBuffer(ctrlInfoToUeId(lteInfo), lteInfo->getLcid());
+
+    UmTxEntity* txbuf = getTxBuffer(lteInfo);
 
     // Create a new RLC packet
     LteRlcSdu* rlcPkt = new LteRlcSdu("rlcUmPkt");
     rlcPkt->setSnoMainPacket(lteInfo->getSequenceNumber());
-    rlcPkt->setByteLength(RLC_HEADER_UM);
+    rlcPkt->setLengthMainPacket(pkt->getByteLength());
     rlcPkt->encapsulate(pkt);
+
+    // create a message so as to notify the MAC layer that the queue contains new data
+    LteRlcPdu* newDataPkt = new LteRlcPdu("newDataPkt");
+    // make a copy of the RLC SDU
+    LteRlcSdu* rlcPktDup = rlcPkt->dup();
+    // the MAC will only be interested in the size of this packet
+    newDataPkt->encapsulate(rlcPktDup);
+    newDataPkt->setControlInfo(lteInfo->dup());
+
+    EV << "LteRlcUm::handleUpperMessage - Sending message " << newDataPkt->getName() << " to port UM_Sap_down$o\n";
+    send(newDataPkt, down_[OUT]);
+
     rlcPkt->setControlInfo(lteInfo);
+
     drop(rlcPkt);
 
-    // Fragment Packet
-    txbuf->fragment(rlcPkt);
+    // Bufferize RLC SDU
+    EV << "LteRlcUm::handleUpperMessage - Enque packet " << rlcPkt->getName() << " into the Tx Buffer\n";
+    txbuf->enque(rlcPkt);
 }
 
 void LteRlcUm::handleLowerMessage(cPacket *pkt)
 {
-    if (strcmp(pkt->getName(), "D2DModeSwitchNotification") == 0)
-    {
-        EV << NOW << " LteRlcUm::handleLowerMessage - Received packet " << pkt->getName() << " from lower layer\n";
-        // add here specific behavior for handling mode switch at the RLC layer
+    EV << "LteRlcUm::handleLowerMessage - Received packet " << pkt->getName() << " from lower layer\n";
 
-        // forward packet to PDCP
-        EV << "LteRlcUm::handleLowerMessage - Sending packet " << pkt->getName() << " to port UM_Sap_up$o\n";
-        send(pkt, up_[OUT]);
+    FlowControlInfo* lteInfo = check_and_cast<FlowControlInfo*>(pkt->getControlInfo());
+
+    if (strcmp(pkt->getName(), "LteMacSduRequest") == 0)
+    {
+        // get the corresponding Tx buffer
+        UmTxEntity* txbuf = getTxBuffer(lteInfo);
+
+        LteMacSduRequest* macSduRequest = check_and_cast<LteMacSduRequest*>(pkt);
+        unsigned int size = macSduRequest->getSduSize();
+
+        drop(pkt);
+
+        // do segmentation/concatenation and send a pdu to the lower layer
+        txbuf->rlcPduMake(size);
+
+        delete macSduRequest;
     }
     else
     {
         // Extract informations from fragment
-        FlowControlInfo* lteInfo = check_and_cast<FlowControlInfo*>(pkt->getControlInfo());
-        MacNodeId nodeId = (lteInfo->getDirection() == DL) ? lteInfo->getDestId() : lteInfo->getSourceId();
-        UmRxQueue* rxbuf = getRxBuffer(nodeId, lteInfo->getLcid());
+        UmRxEntity* rxbuf = getRxBuffer(lteInfo);
         drop(pkt);
 
-        // Defragment packet
-        rxbuf->defragment(pkt);
+        // Bufferize PDU
+        EV << "LteRlcUm::handleLowerMessage - Enque packet " << pkt->getName() << " into the Rx Buffer\n";
+        rxbuf->enque(pkt);
     }
 }
 
 void LteRlcUm::deleteQueues(MacNodeId nodeId)
 {
-    UmTxBuffers::iterator tit;
-    UmRxBuffers::iterator rit;
+    UmTxEntities::iterator tit;
+    UmRxEntities::iterator rit;
 
-    for (tit = txBuffers_.begin(); tit != txBuffers_.end(); )
+    LteNodeType nodeType;
+    std::string nodeTypePar = getAncestorPar("nodeType").stdstringValue();
+    if (strcmp(nodeTypePar.c_str(), "ENODEB") == 0)
+        nodeType = ENODEB;
+    else
+        nodeType = UE;
+
+    // at the UE, delete all connections
+    // at the eNB, delete connections related to the given UE
+    for (tit = txEntities_.begin(); tit != txEntities_.end(); )
     {
-        if (MacCidToNodeId(tit->first) == nodeId)
+        if (nodeType == UE || (nodeType == ENODEB && MacCidToNodeId(tit->first) == nodeId))
         {
-            delete tit->second;           // Delete Queue
-            txBuffers_.erase(tit++);        // Delete Elem
+            delete tit->second;        // Delete Entity
+            txEntities_.erase(tit++);    // Delete Elem
         }
         else
         {
             ++tit;
         }
     }
-    for (rit = rxBuffers_.begin(); rit != rxBuffers_.end(); )
+    for (rit = rxEntities_.begin(); rit != rxEntities_.end(); )
     {
-        if (MacCidToNodeId(rit->first) == nodeId)
+        if (nodeType == UE || (nodeType == ENODEB && MacCidToNodeId(rit->first) == nodeId))
         {
-            delete rit->second;           // Delete Queue
-            rxBuffers_.erase(rit++);        // Delete Elem
+            delete rit->second;        // Delete Entity
+            rxEntities_.erase(rit++);    // Delete Elem
         }
         else
         {
@@ -179,13 +233,28 @@ void LteRlcUm::deleteQueues(MacNodeId nodeId)
 
 void LteRlcUm::initialize()
 {
+    // check the MAC module type: if it is not "realistic", abort simulation
+    std::string nodeType = getParentModule()->getParentModule()->par("nodeType").stdstringValue();
+    std::string macType = getParentModule()->getParentModule()->par("LteMacType").stdstringValue();
+
+    if (nodeType.compare("ENODEB") == 0)
+    {
+        if (macType.compare("LteMacEnbRealistic") != 0)
+            throw cRuntimeError("LteRlcUm::initialize - %s module found, must be LteMacEnbRealistic. Aborting", macType.c_str());
+    }
+    else if (nodeType.compare("UE") == 0)
+    {
+        if (macType.compare("LteMacUeRealistic") != 0)
+            throw cRuntimeError("LteRlcUm::initialize - %s module found, must be LteMacUeRealistic. Aborting", macType.c_str());
+    }
+
     up_[IN] = gate("UM_Sap_up$i");
     up_[OUT] = gate("UM_Sap_up$o");
     down_[IN] = gate("UM_Sap_down$i");
     down_[OUT] = gate("UM_Sap_down$o");
 
-    WATCH_MAP(txBuffers_);
-    WATCH_MAP(rxBuffers_);
+    WATCH_MAP(txEntities_);
+    WATCH_MAP(rxEntities_);
 }
 
 void LteRlcUm::handleMessage(cMessage* msg)
